小功能更新 可以查看

+++ src/com/model/User.java
@@ -4,7 +4,7 @@ package com.model;
 public class User {
     private int td1,td2,td3,td4,td5,td6,td7;
     int tnum = 0;  //通道数
-    int num[];
+    double num[];
     Object NUM2[];
     public int getTd1() {
         return td1;
@@ -49,7 +49,7 @@ public class User {
         this.td7 = t;
     }
 
-    public void setNum(int[] a){
+    public void setNum(double[] a){
         num = a;
         for(int i=0;i<tnum+1;i++)
             NUM2[i] = (Object)num[i];
@@ -60,7 +60,7 @@ public class User {
 
     public void setTnum(int i){
         tnum = i;
-        num = new int[tnum+1]; //数组定义
+        num = new double[tnum+1]; //数组定义
         NUM2 = new Object[tnum+1];  //数组定义
     }
